make search by date compoent reusable
